+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2025-11-03 08:04:14
+
+### Changed Files
+- `src/features/reports/ReportsPage.tsx`
+
+### Statistics
+```
+ 1 file changed, 14 insertions(+), 8 deletions(-)
+```
+
 ## [Checkpoint] - 2025-11-01 14:31:22
 
 ### Changed Files